global UI
This the best that I could do. The problem with testing Tailwind UI is that I cannot determine if it mobile friendly or not without first deploying it as the app relies on parsing a live website. Additionally, many objects and components come with differing top/bottom margins which creates many inconsistency that cannot be fixed without fine-tuning each margin manually.